ref: moved run_test_query to mod level
