+++ src/tests/conformance/query/get_data.rs
@@ -6,32 +6,16 @@
 //!
 //! Note: ZCashd currently ignores requests for non-existant blocks. We expect a [`Message::NotFound`] response.
 
-use std::{io, time::Duration};
-
 use crate::{
     protocol::{
         message::Message,
         payload::{
-            block::Block,
             inv::{InvHash, ObjectKind},
-            Hash, Inv, Nonce,
+            Hash, Inv,
         },
     },
-    setup::node::{Action, Node},
-    tools::synthetic_node::SyntheticNode,
+    tests::conformance::query::{run_test_query, SEED_BLOCKS},
 };
-
-lazy_static::lazy_static!(
-    /// The blocks that the node is seeded with for this test module.
-    static ref SEED_BLOCKS: Vec<Block> = {
-        Block::initial_testnet_blocks()
-    };
-
-    /// InvHashes of the blocks that the node is seeded with.
-    static ref SEED_BLOCK_HASHES: Vec<InvHash> = {
-        SEED_BLOCKS.iter().map(|block| block.inv_hash()).collect()
-    };
-);
 
 mod single_block {
     use super::*;
@@ -42,7 +26,7 @@ mod single_block {
         let block = SEED_BLOCKS.last().unwrap().clone();
         let query = Message::GetData(Inv::new(vec![block.inv_hash()]));
         let expected = vec![Message::Block(block.into())];
-        let response = run_test_case(query).await.unwrap();
+        let response = run_test_query(query).await.unwrap();
         assert_eq!(response, expected);
     }
 
@@ -52,7 +36,7 @@ mod single_block {
         let block = SEED_BLOCKS.first().unwrap().clone();
         let query = Message::GetData(Inv::new(vec![block.inv_hash()]));
         let expected = vec![Message::Block(block.into())];
-        let response = run_test_case(query).await.unwrap();
+        let response = run_test_query(query).await.unwrap();
         assert_eq!(response, expected);
     }
 
@@ -62,7 +46,7 @@ mod single_block {
         let block = SEED_BLOCKS[5].clone();
         let query = Message::GetData(Inv::new(vec![block.inv_hash()]));
         let expected = vec![Message::Block(block.into())];
-        let response = run_test_case(query).await.unwrap();
+        let response = run_test_query(query).await.unwrap();
         assert_eq!(response, expected);
     }
 
@@ -72,7 +56,7 @@ mod single_block {
         let inv = Inv::new(vec![InvHash::new(ObjectKind::Block, Hash::new([17; 32]))]);
         let query = Message::GetData(inv.clone());
         let expected = vec![Message::NotFound(inv)];
-        let response = run_test_case(query).await.unwrap();
+        let response = run_test_query(query).await.unwrap();
         assert_eq!(response, expected);
     }
 }
@@ -90,7 +74,7 @@ mod multiple_blocks {
             .iter()
             .map(|block| Message::Block(Box::new(block.clone())))
             .collect::<Vec<_>>();
-        let response = run_test_case(query).await.unwrap();
+        let response = run_test_query(query).await.unwrap();
         assert_eq!(response, expected);
     }
 
@@ -104,7 +88,7 @@ mod multiple_blocks {
             .iter()
             .map(|&block| Message::Block(Box::new(block.clone())))
             .collect::<Vec<_>>();
-        let response = run_test_case(query).await.unwrap();
+        let response = run_test_query(query).await.unwrap();
         assert_eq!(response, expected);
     }
 
@@ -118,7 +102,7 @@ mod multiple_blocks {
             .iter()
             .map(|block| Message::Block(Box::new(block.clone())))
             .collect::<Vec<_>>();
-        let response = run_test_case(query).await.unwrap();
+        let response = run_test_query(query).await.unwrap();
         assert_eq!(response, expected);
     }
 
@@ -133,7 +117,7 @@ mod multiple_blocks {
 
         let query = Message::GetData(inv.clone());
         let expected = vec![Message::NotFound(inv)];
-        let response = run_test_case(query).await.unwrap();
+        let response = run_test_query(query).await.unwrap();
         assert_eq!(response, expected);
     }
 
@@ -175,7 +159,7 @@ mod multiple_blocks {
             .collect();
 
         let query = Message::GetData(Inv::new(mixed_inv));
-        let response = run_test_case(query).await.unwrap();
+        let response = run_test_query(query).await.unwrap();
 
         // Should contain expected_blocks[..] and expected_non_existant. Not sure
         // what order we should expect (if any). But since the node currently ignores
@@ -184,52 +168,4 @@ mod multiple_blocks {
         expected.push(expected_non_existant);
         assert_eq!(response, expected);
     }
-}
-
-/// Starts a node seeded with the initial testnet chain, connects a single
-/// SyntheticNode and sends a query. The node's responses to this query is
-/// then returned.
-async fn run_test_case(query: Message) -> io::Result<Vec<Message>> {
-    // Spin up a node instance with knowledge of the initial testnet-chain.
-    let mut node = Node::new().unwrap();
-    node.initial_action(Action::SeedWithTestnetBlocks(SEED_BLOCKS.len()))
-        .start()
-        .await?;
-
-    // Create a synthetic node.
-    let mut synthetic_node = SyntheticNode::builder()
-        .with_full_handshake()
-        .with_all_auto_reply()
-        .build()
-        .await?;
-
-    // Connect to the node and initiate handshake.
-    synthetic_node.connect(node.addr()).await?;
-
-    // Send the query.
-    synthetic_node
-        .send_direct_message(node.addr(), query)
-        .await?;
-
-    // Send a Ping - once we receive the matching Pong we know our query has been fully processed.
-    let nonce = Nonce::default();
-    synthetic_node
-        .send_direct_message(node.addr(), Message::Ping(nonce))
-        .await?;
-
-    // Receive messages until we receive the matching Pong, or we timeout.
-    const RECV_TIMEOUT: Duration = Duration::from_millis(100);
-    let mut messages = Vec::new();
-    loop {
-        match synthetic_node.recv_message_timeout(RECV_TIMEOUT).await? {
-            (_, Message::Pong(rx_nonce)) if rx_nonce == nonce => break,
-            (_, message) => messages.push(message),
-        }
-    }
-
-    // Gracefully shut down the nodes.
-    synthetic_node.shut_down();
-    node.stop()?;
-
-    Ok(messages)
 }
